feat: 综合更新 - MallCenter优化 + 前后端修复 + 文档清理
MallCenter 小程序:
- 修复 WXSS 组件选择器错误 (goods-card)
- 替换 emoji 为纯中文文字
- 修复 wx:key 警告
- 优化商品列表和详情页样式

MallCenter 后台:
- 更新 API 配置和安全设置
- 优化订单和商品管理

React Native 前端:
- 修复供应商/客户 API 调用 (params 修复)
- 字段名对齐后端 (phone, shippingAddress)
- 优化管理界面

后端 Java:
- 新增 Controller (Disposal, Shipment, Traceability 等)
- 完善 DTO 和 Repository
- AI 分析服务优化

文档清理:
- 移除过期报告和测试文件
- 整理 PRD 文档结构

+++ backend-java/src/main/java/com/cretas/aims/controller/UserController.java
@@ -63,7 +63,7 @@ public class UserController {
             @Parameter(description = "工厂ID", required = true)
             @PathVariable @NotBlank String factoryId,
             @Parameter(description = "用户ID", required = true)
-            @PathVariable @NotNull Integer userId,
+            @PathVariable @NotNull Long userId,
             @Valid @RequestBody CreateUserRequest request) {
         log.info("更新用户: factoryId={}, userId={}", factoryId, userId);
         UserDTO user = userService.updateUser(factoryId, userId, request);
@@ -79,7 +79,7 @@ public class UserController {
             @Parameter(description = "工厂ID", required = true)
             @PathVariable @NotBlank String factoryId,
             @Parameter(description = "用户ID", required = true)
-            @PathVariable @NotNull Integer userId) {
+            @PathVariable @NotNull Long userId) {
         log.info("删除用户: factoryId={}, userId={}", factoryId, userId);
         userService.deleteUser(factoryId, userId);
         return ApiResponse.success("用户删除成功", null);
@@ -94,7 +94,7 @@ public class UserController {
             @Parameter(description = "工厂ID", required = true)
             @PathVariable @NotBlank String factoryId,
             @Parameter(description = "用户ID", required = true)
-            @PathVariable @NotNull Integer userId) {
+            @PathVariable @NotNull Long userId) {
         UserDTO user = userService.getUserById(factoryId, userId);
         return ApiResponse.success(user);
     }
@@ -135,7 +135,7 @@ public class UserController {
             @Parameter(description = "工厂ID", required = true)
             @PathVariable @NotBlank String factoryId,
             @Parameter(description = "用户ID", required = true)
-            @PathVariable @NotNull Integer userId) {
+            @PathVariable @NotNull Long userId) {
         log.info("激活用户: factoryId={}, userId={}", factoryId, userId);
         userService.activateUser(factoryId, userId);
         return ApiResponse.success("用户激活成功", null);
@@ -150,7 +150,7 @@ public class UserController {
             @Parameter(description = "工厂ID", required = true)
             @PathVariable @NotBlank String factoryId,
             @Parameter(description = "用户ID", required = true)
-            @PathVariable @NotNull Integer userId) {
+            @PathVariable @NotNull Long userId) {
         log.info("停用用户: factoryId={}, userId={}", factoryId, userId);
         userService.deactivateUser(factoryId, userId);
         return ApiResponse.success("用户停用成功", null);
@@ -165,7 +165,7 @@ public class UserController {
             @Parameter(description = "工厂ID", required = true)
             @PathVariable @NotBlank String factoryId,
             @Parameter(description = "用户ID", required = true)
-            @PathVariable @NotNull Integer userId,
+            @PathVariable @NotNull Long userId,
             @Parameter(description = "新角色", required = true)
             @RequestParam FactoryUserRole newRole) {
         log.info("更新用户角色: factoryId={}, userId={}, newRole={}", factoryId, userId, newRole);
